requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
autoflake==2.3.1
black==25.1.0
cfgv==3.4.0
click==8.2.1
distlib==0.3.9
factory_boy==3.3.3
Faker==37.4.0
fastapi==0.115.14
filelock==3.18.0
flake8==7.3.0
flake8-docstrings==1.7.0
Flake8-pyproject==1.2.3
greenlet==3.2.3
h11==0.16.0
identify==2.6.12
idna==3.10
isort==6.0.1
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
mypy==1.16.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pre_commit==4.2.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydocstyle==6.3.0
pyflakes==3.4.0
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
virtualenv==20.31.2
